+++ Casa/src/servicios/ServiciosCocina.java
@@ -96,7 +96,7 @@ public class ServiciosCocina implements Observer
             }
         }
         this.kernel.setRecetas(recetas);
-        this.kernel.setIntesidadLuz((int)this.perfil.getIntesidadLuz());
+        this.kernel.setIntesidadLuz(this.perfil.getIntesidadLuz());
     }
 
     private Contenedor getContenedor(String nombre)
